Reorganized according to example project in moodle (V2).
