+++ Entities/Entities/Block.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Raktar;
+namespace Raktar.DataContext.Entities;
 
 public partial class Block
 {
