+++ Raktar.Services/Raktar.Services.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Entities\Raktar.DataContext.csproj" />
+  </ItemGroup>
+
 </Project>
