+++ Raktar.sln
@@ -5,9 +5,9 @@ VisualStudioVersion = 17.13.35806.99
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raktar", "Raktar\Raktar.csproj", "{A3B223B6-5B90-4E8D-B132-0C2DA48292B3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Entities", "Entities\Entities.csproj", "{3354210D-EC10-4F61-8289-C9C55F6CD983}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raktar.DataContext", "Entities\Raktar.DataContext.csproj", "{3354210D-EC10-4F61-8289-C9C55F6CD983}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Controllers", "Controllers\Controllers.csproj", "{A2A7868A-9F4E-4503-9A4A-C5A24698C200}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raktar.Services", "Raktar.Services\Raktar.Services.csproj", "{963843C7-BE65-4CA5-A4FE-9E10C054A4B2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Release|Any CPU.Build.0 = Release|Any CPU
+		{963843C7-BE65-4CA5-A4FE-9E10C054A4B2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{963843C7-BE65-4CA5-A4FE-9E10C054A4B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{963843C7-BE65-4CA5-A4FE-9E10C054A4B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{963843C7-BE65-4CA5-A4FE-9E10C054A4B2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
